+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod env;
 pub mod executer;
 pub mod global_flag_storage;
 pub mod logger;
+pub mod placeholder_storage;
 pub mod types;
 pub mod utils;