+++ src/types/placeholder_definition.rs
@@ -1,0 +1,5 @@
+pub(crate) struct PlaceholderDefinition {
+    pub name: String,
+    pub description: String,
+    pub get_value: fn() -> String,
+}