+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [v2.0.0](https://github.com/cophilot/templify/tree/2.0.0) (2025-X-X) _IN DEVELOPMENT_
 
+-   Bug fixes [#18](https://github.com/cophilot/templify/issues/18)
 -   Using YAML format for `.templify` (now `.templify.yml`) files [#5](https://github.com/cophilot/templify/issues/5)
 -   Generate snippets when generatng from a template [#3](https://github.com/cophilot/templify/issues/3)
 -   Added `placeholder` command [#21](https://github.com/cophilot/templify/issues/21)